research_engine/pyproject.toml
[project]
name = "web-scraper"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.9"
dependencies = [
    "bs4>=0.0.2",
    "ddgs>=9.8.0",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "ollama>=0.6.1",
    "python-dotenv>=1.2.1",
    "readability-lxml>=0.8.4.1",
    "requests>=2.32.5",
    "selenium>=4.36.0",
    "uvicorn>=0.39.0",
]
